backend/requirements.txt
flask
flask-cors
opencv-python
opencv-contrib-python
numpy
pandas
pillow
